Moved some files to Components package
