+++ app/src/main/java/com/belia/speedotransfer/ui/auth/SecondSignUpScreen.kt
@@ -30,7 +30,7 @@ import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import com.belia.speedotransfer.ui.auth.components.CountryPicker
 import com.belia.speedotransfer.ui.auth.components.DatePicker
-import com.belia.speedotransfer.ui.common_ui.TopBar
+import com.belia.speedotransfer.ui.components.TopBar
 import com.belia.speedotransfer.ui.theme.BottomRose
 import com.belia.speedotransfer.ui.theme.GrayG0
 import com.belia.speedotransfer.ui.theme.GrayG100
