+++ app/src/main/java/com/belia/speedotransfer/ui/auth/components/TopBar.kt
@@ -1,10 +1,9 @@
-package com.belia.speedotransfer.ui.common_ui
+package com.belia.speedotransfer.ui.components
 
 import androidx.compose.material3.CenterAlignedTopAppBar
 import androidx.compose.material3.ExperimentalMaterial3Api
 import androidx.compose.material3.Icon
 import androidx.compose.material3.IconButton
-import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Text
 import androidx.compose.material3.TopAppBarDefaults
 import androidx.compose.runtime.Composable
@@ -12,8 +11,8 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.tooling.preview.Preview
-import androidx.compose.ui.unit.sp
 import com.belia.speedotransfer.R
+import com.belia.speedotransfer.ui.theme.titleMedium
 
 @OptIn(ExperimentalMaterial3Api::class)
 @Composable
@@ -27,13 +26,12 @@ fun TopBar(
         title = {
             Text(
                 text = title,
-                style = MaterialTheme.typography.titleMedium,
-                fontSize = 20.sp,
+                style = titleMedium,
             )
         },
         navigationIcon = {
             if (hasIcon) {
-                IconButton(onClick = { /*TODO*/ }) {
+                IconButton(onClick = { /*TODO(popBackStack)*/ }) {
                     Icon(
                         painter = painterResource(id = R.drawable.ic_back),
                         contentDescription = "back button"
@@ -45,10 +43,4 @@ fun TopBar(
             containerColor = color
         ),
     )
-}
-
-@Preview
-@Composable
-private fun TopBarPrev() {
-    TopBar(color = Color(0xFFFFF8E7), hasIcon = true, title = "Transfer")
 }
